Switch to h2 for cv#2

+++ src/main/java/PennyPincher/service/expenses/ExpenseServiceImpl.java
@@ -1,32 +1,33 @@
 package PennyPincher.service.expenses;
 
+import PennyPincher.dto.expense.CustomExpenseDto;
+import PennyPincher.dto.expense.ExpenseMapper;
+import PennyPincher.entity.Event;
 import PennyPincher.entity.Expense;
 import PennyPincher.entity.Payoff;
 import PennyPincher.entity.User;
-import PennyPincher.repository.EventRepository;
+import PennyPincher.exception.ExpenseNotFoundException;
 import PennyPincher.repository.ExpenseRepository;
-import PennyPincher.repository.UserRepository;
+import PennyPincher.service.users.UserService;
 import lombok.AllArgsConstructor;
-import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Service
 public class ExpenseServiceImpl implements ExpenseService {
-    private ExpenseRepository expenseRepository;
+
+    private final ExpenseRepository expenseRepository;
+    private final UserService userService;
 
     @Override
     public Expense findById(Integer expenseId) {
         return expenseRepository.findById(expenseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Expense not found with id: " + expenseId));
+                .orElseThrow(() -> new ExpenseNotFoundException("Expense not found with ID: " + expenseId));
     }
 
     @Override
@@ -35,15 +36,26 @@ public class ExpenseServiceImpl implements ExpenseService {
     }
 
     @Override
+    public void deleteByEventId(Integer eventId) {
+        expenseRepository.deleteById(eventId);
+    }
+
+    @Override
     public void deleteById(Integer expenseId) {
+        if (!expenseRepository.existsById(expenseId)) {
+            throw new ExpenseNotFoundException("Expense not found with ID: " + expenseId);
+        }
         expenseRepository.deleteById(expenseId);
     }
 
     @Override
     public List<Expense> findExpensesForGivenEvent(Integer eventId) {
-        return expenseRepository.findAll().stream()
-                .filter(expense -> eventId.equals(expense.getEvent().getId()))
-                .collect(Collectors.toList());
+        return expenseRepository.findByEventId(eventId);
+    }
+
+    @Override
+    public Expense findByExpenseNameAndEventId(String expenseName, Integer eventId) {
+        return expenseRepository.findByNameAndEventId(expenseName, eventId);
     }
 
     @Override
@@ -55,39 +67,87 @@ public class ExpenseServiceImpl implements ExpenseService {
 
     @Override
     public Map<Integer, BigDecimal> mapUserToCost(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerCost = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerCost.put(participant.getId(),
-                sumParticipantCosts(expense, participant)
-        ));
-        return mapUserPerCost;
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> sumParticipantCosts(expense, participant)
+                ));
     }
 
     @Override
     public Map<Integer, BigDecimal> mapUserToPayoffAmount(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerPayoffsAmount = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerPayoffsAmount.put(participant.getId(),
-                sumParticipantPayoffs(expense, participant)
-        ));
-        return mapUserPerPayoffsAmount;
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> sumParticipantPayoffs(expense, participant)
+                ));
     }
 
     @Override
     public Map<Integer, BigDecimal> mapUserToBalance(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerBalance = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerBalance.put(participant.getId(),
-                calculateParticipantBalance(expense, participant)
-        ));
-        return mapUserPerBalance;
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> calculateParticipantBalance(expense, participant)
+                ));
     }
 
     @Override
-    public void deleteByEventId(Integer eventId) {
-        expenseRepository.deleteById(eventId);
+    public void updateExpenseAttributes(List<Expense> eventExpenses) {
+        for (Expense expense : eventExpenses) {
+            Map<Integer, BigDecimal> payoffAmountPerParticipant = mapUserToPayoffAmount(expense);
+            Map<Integer, BigDecimal> balancePerParticipant = mapUserToBalance(expense);
+            Map<Integer, BigDecimal> costPerParticipant = mapUserToCost(expense);
+            expense.setCostPerUser(costPerParticipant);
+            expense.setPayoffPerUser(payoffAmountPerParticipant);
+            expense.setBalancePerUser(balancePerParticipant);
+            save(expense);
+        }
     }
 
     @Override
-    public Expense findByExpenseNameAndEventId(String expenseName, Integer eventId) {
-        return expenseRepository.findByNameAndEventId(expenseName, eventId);
+    public BigDecimal calculateUpdatedBalanceForEvent(List<Expense> eventExpenses) {
+        return eventExpenses.stream()
+                .flatMap(expense -> expense.getBalancePerUser().values().stream())
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
+    }
+
+    @Override
+    public Expense createExpense(Event foundEvent,
+                                 User loggedInUser,
+                                 CustomExpenseDto customExpenseDto,
+                                 ExpenseMapper expenseMapper) {
+        List<User> eventMembers = foundEvent.getEventMembers();
+        List<String> namesOfMembers = eventMembers.stream()
+                .map(User::getUsername)
+                .collect(Collectors.toList());
+        customExpenseDto.setParticipantsNames(namesOfMembers);
+
+        Expense expense = expenseMapper.mapCustomExpenseDtoToDomain(foundEvent, customExpenseDto);
+
+        List<Payoff> expensePayoffs = new ArrayList<>();
+        Payoff defaultPayoff = Payoff.builder()
+                .expensePaid(expense)
+                .userPaying(loggedInUser)
+                .payoffAmount(BigDecimal.ZERO)
+                .build();
+        expensePayoffs.add(defaultPayoff);
+        expense.setPayoffs(expensePayoffs);
+
+        return expense;
+    }
+
+    @Override
+    public void updateParticipantsAndDeleteExpense(Expense expense,
+                                                   Map<Integer, BigDecimal> costPerParticipant,
+                                                   Map<Integer, BigDecimal> payoffPerParticipant) {
+        expense.getParticipants().forEach(participant -> {
+            BigDecimal participantBalanceChange = costPerParticipant.getOrDefault(participant.getId(), BigDecimal.ZERO)
+                    .subtract(payoffPerParticipant.getOrDefault(participant.getId(), BigDecimal.ZERO));
+            participant.setBalance(participant.getBalance().add(participantBalanceChange));
+            participant.getExpenses().removeIf(exp -> exp.getId().equals(expense.getId()));
+            userService.save(participant);
+        });
     }
 
     private BigDecimal sumParticipantPayoffs(Expense expense, User participant) {
